# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen rental PlayStation Byone Arena. Mendukung realtime via WebSocket untuk aplikasi mobile dan TV Android.",
        "title": "Byone Arena API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Byone Arena Support",
            "email": "support@byone-arena.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/discounts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan seluruh aturan diskon otomatis (happy hour, member, dll) beserta statusnya",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Ambil semua aturan diskon",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.DiscountRule"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat aturan diskon otomatis baru. Tipe rule: 'always', 'happy_hour', 'member', 'day_of_week'.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Buat aturan diskon baru",
                "parameters": [
                    {
                        "description": "Data aturan diskon baru",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateDiscountRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.DiscountRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/discounts/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan aturan diskon yang sedang aktif, diurutkan berdasarkan prioritas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Ambil aturan diskon aktif",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.DiscountRule"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/discounts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan detail satu aturan diskon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Ambil aturan diskon berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discount Rule ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.DiscountRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah data aturan diskon. Semua field opsional (partial update).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Update aturan diskon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discount Rule ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data yang diubah",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.UpdateDiscountRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.DiscountRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus aturan diskon secara permanen",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Hapus aturan diskon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discount Rule ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/discounts/{id}/toggle": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Toggle status aktif aturan diskon (aktif ↔ nonaktif)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Aktifkan / nonaktifkan aturan diskon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discount Rule ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.DiscountRule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.DiscountRule": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "daysOfWeek": {
                    "description": "Day of week: comma-separated \"0,1,2\" (0=Minggu, 1=Senin, ..., 6=Sabtu)",
                    "type": "string"
                },
                "discountType": {
                    "description": "reuse dari voucher.go",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.DiscountType"
                        }
                    ]
                },
                "discountValue": {
                    "type": "number"
                },
                "endHour": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "maxDiscount": {
                    "description": "batas maks diskon persen (0 = tidak terbatas)",
                    "type": "number"
                },
                "minPurchase": {
                    "description": "minimal total sebelum rule berlaku",
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "description": "lebih besar = lebih dulu dievaluasi",
                    "type": "integer"
                },
                "ruleType": {
                    "$ref": "#/definitions/entity.RuleType"
                },
                "startHour": {
                    "description": "Happy hour: jam mulai dan jam selesai (0-23, bisa lintas tengah malam)",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.DiscountType": {
            "type": "string",
            "enum": [
                "percentage",
                "fixed_amount"
            ],
            "x-enum-comments": {
                "DiscountTypeFixedAmount": "diskon nominal tetap",
                "DiscountTypePercentage": "diskon persen dari total"
            },
            "x-enum-varnames": [
                "DiscountTypePercentage",
                "DiscountTypeFixedAmount"
            ]
        },
...
        "entity.RuleType": {
            "type": "string",
            "enum": [
                "always",
                "happy_hour",
                "member",
                "day_of_week"
            ],
            "x-enum-comments": {
                "RuleTypeAlways": "berlaku untuk semua transaksi",
                "RuleTypeDayOfWeek": "berlaku pada hari tertentu",
                "RuleTypeHappyHour": "berlaku pada jam tertentu",
                "RuleTypeMember": "khusus pelanggan member"
            },
            "x-enum-varnames": [
                "RuleTypeAlways",
                "RuleTypeHappyHour",
                "RuleTypeMember",
                "RuleTypeDayOfWeek"
            ]
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "usecase.CreateDiscountRuleRequest": {
            "type": "object",
            "required": [
                "discountType",
                "discountValue",
                "name",
                "ruleType"
            ],
            "properties": {
                "daysOfWeek": {
                    "description": "wajib jika rule_type=day_of_week, format \"0,1,2\"",
                    "type": "string"
                },
                "discountType": {
                    "enum": [
                        "percentage",
                        "fixed_amount"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.DiscountType"
                        }
                    ]
                },
                "discountValue": {
                    "type": "number"
                },
                "endHour": {
                    "description": "wajib jika rule_type=happy_hour (0-23)",
                    "type": "integer"
                },
                "maxDiscount": {
                    "type": "number",
                    "minimum": 0
                },
                "minPurchase": {
                    "type": "number",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 3
                },
                "priority": {
                    "type": "integer",
                    "minimum": 0
                },
                "ruleType": {
                    "enum": [
                        "always",
                        "happy_hour",
                        "member",
                        "day_of_week"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.RuleType"
                        }
                    ]
                },
                "startHour": {
                    "description": "wajib jika rule_type=happy_hour (0-23)",
                    "type": "integer"
                }
            }
        },
...
        "usecase.UpdateDiscountRuleRequest": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "string"
                },
                "discountType": {
                    "enum": [
                        "percentage",
                        "fixed_amount"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.DiscountType"
                        }
                    ]
                },
                "discountValue": {
                    "type": "number"
                },
                "endHour": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "maxDiscount": {
                    "type": "number",
                    "minimum": 0
                },
                "minPurchase": {
                    "type": "number",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 3
                },
                "priority": {
                    "type": "integer",
                    "minimum": 0
                },
                "ruleType": {
                    "enum": [
                        "always",
                        "happy_hour",
                        "member",
                        "day_of_week"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.RuleType"
                        }
                    ]
                },
                "startHour": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}